refactor: Use context to store user
