+++ client/src/helpers/helpers.ts
@@ -1,3 +1,4 @@
+
 export const setLocalStorage = (key, value) => {
   if (window !== "undefined") {
     localStorage.setItem(key, JSON.stringify(value));
